# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Wardrobe App API
  version: 0.1.0
  description: |
    Wardrobe App のAPI仕様。

    このAPIは Cookie ベースのセッション認証を使用する。
    CSRF 保護が必要な POST / PUT / DELETE では X-CSRF-TOKEN ヘッダが必要。
    docs 上では認証系 API / 設定系 API / 参照マスタ API を軸に整理しつつ、Items / Outfits も実装単位で記載する。
    `400` / `429` / `500` は、アプリ独自の契約がある場合を除き framework default に委ね、通常の endpoint 定義では過剰に明示しない。
    `care_status` のような補助状態は、主 status と別カラムで扱い、候補除外や invalid 化の主制御には使わない。
servers:
  - url: http://localhost:8000
    description: ローカル開発環境
...
paths:
...
  /api/weather-records:
    get:
      tags:
        - Wear Logs
      summary: 指定日の天気記録一覧を取得
      operationId: getWeatherRecordsByDate
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: 天気記録一覧
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeatherRecordsResponse"
        "422":
          description: バリデーションエラー
    post:
      tags:
        - Wear Logs
      summary: 天気記録を作成
      operationId: createWeatherRecord
      parameters:
        - $ref: "#/components/parameters/XCsrfTokenHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeatherRecordUpsertRequest"
      responses:
        "201":
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeatherRecordMutationResponse"
        "422":
          description: バリデーションエラー
...
components:
  parameters:
    XCsrfTokenHeader:
      name: X-CSRF-TOKEN
      in: header
      required: false
      description: |
        CSRF対策用ヘッダ。
        /csrf-cookie 実行後に取得した XSRF-TOKEN と同じ値を指定する。
        実運用では必須。
        Swagger UI では requestInterceptor により自動付与する。
      schema:
        type: string
...
  schemas:
...
    MessageOnlyResponse:
      description: |
        action 完了を表す最小 response。
        `message` には `deleted` / `logged_out` などの action keyword を入れる。
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: logged_out
...
    WeatherCondition:
      type: string
      enum:
        - sunny
        - cloudy
        - rain
        - snow
        - other
    WeatherSourceType:
      type: string
      enum:
        - manual
        - forecast_api
        - historical_api
...
    WeatherRecord:
      type: object
      required:
        - id
        - weather_date
        - location_id
        - location_name
        - location_name_snapshot
        - weather_condition
        - source_type
        - source_name
      properties:
        id:
          type: integer
        weather_date:
          type: string
          format: date
        location_id:
          type: integer
          nullable: true
        location_name:
          type: string
        location_name_snapshot:
          type: string
        forecast_area_code_snapshot:
          type: string
          nullable: true
          description: 保存時点の予報API用地域コード snapshot。MVPでは weather.tsukumijima.net の city code を主対象とする。
        weather_condition:
          $ref: "#/components/schemas/WeatherCondition"
        temperature_high:
          type: number
          nullable: true
        temperature_low:
          type: number
          nullable: true
        memo:
          type: string
          nullable: true
        source_type:
          $ref: "#/components/schemas/WeatherSourceType"
        source_name:
          type: string
        source_fetched_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
    WeatherRecordsResponse:
      type: object
      required:
        - weatherRecords
      properties:
        weatherRecords:
          type: array
          items:
            $ref: "#/components/schemas/WeatherRecord"
    WeatherRecordUpsertRequest:
      type: object
      required:
        - weather_date
        - weather_condition
      properties:
        weather_date:
          type: string
          format: date
        location_id:
          type: integer
          nullable: true
        location_name:
          type: string
          nullable: true
        save_location:
          type: boolean
        weather_condition:
          $ref: "#/components/schemas/WeatherCondition"
        temperature_high:
          type: number
          nullable: true
        temperature_low:
          type: number
          nullable: true
        memo:
          type: string
          nullable: true
    WeatherRecordMutationResponse:
      allOf:
        - $ref: "#/components/schemas/MessageOnlyResponse"
        - type: object
          required:
            - weatherRecord
          properties:
            weatherRecord:
              $ref: "#/components/schemas/WeatherRecord"
  securitySchemes:
    SessionCookieAuth:
      type: apiKey
      in: cookie
      name: laravel-session